feat: move mwst as stores connection, wallet object into separate file
