+++ acapy_wallet_upgrade/__main__.py
@@ -7,9 +7,10 @@ import sys
 from typing import Dict, Optional
 from urllib.parse import urlparse
 
-from .error import UpgradeError, MissingWalletError
+from .error import UpgradeError
 from .pg_connection import PgConnection
-from .pg_mwst_connection import PgMWSTConnection, PgMWSTStoresConnection
+from .pg_mwst_connection import PgMWSTConnection
+from .pg_mwst_stores_connection import PgMWSTStoresConnection
 from .sqlite_connection import SqliteConnection
 from .strategies import DbpwStrategy, MwstAsProfilesStrategy, MwstAsStoresStrategy
